Add vector resources
