+++ app/src/main/res/drawable/flag_australia.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M0,0h512v512H0z"
+      android:fillColor="#00008B"/>
+  <path
+      android:pathData="M256,0v32l-95,96 95,93.5V256h-33.5L127,162l-93,94H0v-34l93,-93.5L0,37V0h31l96,94 93,-94z"
+      android:fillColor="#fff"/>
+  <path
+      android:pathData="m92,162 l5.5,17L21,256L0,256v-1.5zM154,156 L181,160 256,233.5L256,256zM256,0l-96,98 -2,-22 75,-76zM0,0.5 L96.5,95 67,91 0,24.5z"
+      android:fillColor="#ff0000"/>
+  <path
+      android:pathData="M88,0v256h80V0zM0,88v80h256V88z"
+      android:fillColor="#fff"/>
+  <path
+      android:pathData="M0,104v48h256v-48zM104,0v256h48V0z"
+      android:fillColor="#ff0000"/>
+  <path
+      android:pathData="m202,402.8 l-45.8,5.4 4.6,45.9 -32.8,-32.4 -33,32.2 4.9,-45.9 -45.8,-5.8L93,377.4 69,338l43.6,15 15.8,-43.4 15.5,43.5 43.7,-14.7 -24.3,39.2 38.8,25.1ZM424.7,410.8 L404.2,413.4 406.4,433.9 391.6,419.5 376.9,434 378.9,413.5 358.4,411.1 375.7,399.9 364.8,382.4 384.4,388.9 391.3,369.4 398.4,388.8 417.9,382.1 407.2,399.7zM415,293.6l2.7,-13 -9.8,-9 13.2,-1.5 5.5,-12.1 5.5,12.1 13.2,1.5 -9.8,9 2.7,13 -11.6,-6.6zM330.9,233.6 L310.6,235.8 312.4,256.1 298,241.6 283.2,255.7 285.6,235.4 265.4,232.7 282.7,221.9 272.2,204.4 291.5,211.2 298.7,192.1 305.4,211.4 324.8,205.1 313.9,222.4zM506.7,200.8 L485.8,203.5 488.1,224.4 473,209.7 458,224.5 460.1,203.5 439.2,201.1 456.9,189.6 445.8,171.7 465.8,178.4 472.8,158.6 480,178.4 499.9,171.5 488.9,189.5zM424.6,117.3 L403.9,119.6 405.8,140.4 391.1,125.6L376,140l2.4,-20.7 -20.7,-2.8 17.7,-11 -10.7,-17.9 19.7,6.9 7.3,-19.5 6.8,19.7 19.8,-6.5 -11.1,17.6z"
+      android:fillColor="#fff"/>
+</vector>
