+++ app/src/main/res/drawable/flag_france.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M0,0h512v512H0z"
+      android:fillColor="#fff"/>
+  <path
+      android:pathData="M0,0h170.7v512H0z"
+      android:fillColor="#000091"/>
+  <path
+      android:pathData="M341.3,0H512v512H341.3z"
+      android:fillColor="#e1000f"/>
+</vector>
