+++ app/src/main/res/drawable/flag_germany.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M0,341.3h512V512H0z"
+      android:fillColor="#ffce00"/>
+  <path
+      android:pathData="M0,0h512v170.7H0z"
+      android:fillColor="#000001"/>
+  <path
+      android:pathData="M0,170.7h512v170.6H0z"
+      android:fillColor="#ff0000"/>
+</vector>
