+++ app/src/main/res/drawable/flag_usa.xml
@@ -1,0 +1,20 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M0,0h512v512H0"
+      android:fillColor="#bd3d44"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M0,58h512M0,137h512M0,216h512M0,295h512M0,374h512M0,453h512"
+      android:strokeWidth="40"
+      android:strokeColor="#fff"/>
+  <path
+      android:pathData="M0,0h390v275H0z"
+      android:fillColor="#192f5d"/>
+  <path
+      android:pathData="m15,0 l9.3,28.6L0,11h30L5.7,28.6"
+      android:fillColor="#fff"/>
+</vector>
